pyproject.toml
[project]
name = "generation-scaling"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "jaydrennan", email = "jsdman1313@gmail.com" }
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.28.0",
    "ruff>=0.3.3",
    "multipart>=0.2.4",
    "python-multipart>=0.0.9",
    "modal>=0.61.58",
    "imaginairy>=14.1.1",
    "pillow>=10.2.0",
    "pydantic>=2.6.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/generation_scaling"]
